Moved main into a new file and corrected build.gradle
